Moved messages package to domain
